[refactor][coordinator] Refactor coordinator_client and add GetTableMetrics, GetStoreMetrics function calls.
